コンボの新規生成、子を生成してその編集メニューに移るまでを作成

+++ Class/Condition/child/Keyboard/PressKeyboardCondition.cpp
@@ -3,5 +3,5 @@
 bool LWP::Utility::PressKeyboardCondition::CheckCondition()
 {
 	// キーの長押し入力を検知してそれを返す
-	return LWP::Input::Keyboard::GetPress(kId_);
+	return LWP::Input::Keyboard::GetPress(static_cast<uint8_t>(kId_));
 }
